Address Gemini Code Assist comments about int64's

+++ tensorflow_quantum/core/ops/math_ops/tfq_inner_product.cc
@@ -226,7 +226,7 @@ class TfqInnerProductOp : public tensorflow::OpKernel {
 
     const int output_dim_internal_size = output_tensor->dimension(1);
 
-    auto DoWork = [&](int start, int end) {
+    auto DoWork = [&](int64_t start, int64_t end) {
       int old_batch_index = -2;
       int cur_batch_index = -1;
       uint64_t largest_nq = 1;
@@ -236,7 +236,7 @@ class TfqInnerProductOp : public tensorflow::OpKernel {
       StateSpace ss = StateSpace(tfq_for);
       auto sv = ss.Create(largest_nq);
       auto scratch = ss.Create(largest_nq);
-      for (int i = start; i < end; i++) {
+      for (int64_t i = start; i < end; i++) {
         cur_batch_index = i / output_dim_internal_size;
         cur_internal_index = i % output_dim_internal_size;
